vertical alignment
since we're big leaguin now, we have entered the two digit level of stars. while managing the vast wealth that has come from this, a new issue cropped up.. several digit star levels offset the vertical alignment of the rest, lesser, more common, inferior projects with single-digit, mere-mortal levels of popularity, number of stars. this should be fixed. so now we put a cap of 3 on there, meaning that if we have something more than the hundreds, we're going to give it a human readable suffix like 1.2k or 2.3M.